Adds code and assets from previous projects.
I already had some code from a similar game I had been working on.
I iported it so it can be useful in the project.

Adds some extra classes for scriptable objects and others to allow
the creation and configuration of outfits and their required info.

Sprites, prefabs, animations and tilesets are also imported from the
previous project, to provide a solid ground to start the task.