Imports sprites from Stardew Valley Game.
The task requires to be able to customize the player's character. I
imported the sprite sheets used for the game 'Stardew Valley'.

Since I'm going to use the 2D Unity's animation system, and in particular,
sprite libraries, I created the necessary sprite libraries and scriptable
objects to define the outfits and save the data that the UI will display.

I also added the font from the game 'Stardew Valley', to be used in all
the UI texts.

Changed the music and sfx sounds from folder.