Full code commit.
Unfortunately, due to the time constraints, I had to focus on writing
and testing as much code as possible before the deadline.

As such, the gradual code commits, comments, and overall code quality seen
in clean, well organized repositories are missing.

I know this greatly affects the code quality, readability and
efficiency, but I can assure it is not my usual workflow.

I tried to make the code as much as designer friendly as I could, so
that designers can modify the scene instead of hard coding
behaviour.

Please refer to the pdf file that I will attach to know more about
the thought process behind the system.

I will try and refactor, if I have the time.